add comments in aggregateexecutor; separate the declaration of UserDefinedFunctionRunner, UserDefinedAggregateFunctionRunner, and UserDefinedScalarFunctionRunner to different java files
