+++ src/frontend/org/voltdb/jni/ExecutionEngineJNI.java
@@ -41,8 +41,8 @@ import org.voltdb.SnapshotCompletionMonitor.ExportSnapshotTuple;
 import org.voltdb.StatsSelector;
 import org.voltdb.TableStreamType;
 import org.voltdb.TheHashinator.HashinatorConfig;
-import org.voltdb.UserDefinedFunctionManager.UserDefinedScalarFunctionRunner;
-import org.voltdb.UserDefinedFunctionManager.UserDefinedAggregateFunctionRunner;
+import org.voltdb.UserDefinedScalarFunctionRunner;
+import org.voltdb.UserDefinedAggregateFunctionRunner;
 import org.voltdb.VoltDB;
 import org.voltdb.VoltTable;
 import org.voltdb.VoltType;
